+++ src/pages/login/constants.ts
@@ -1,4 +1,17 @@
 export const LOGIN = {
   errorsWrapper: ['w-full', 'max-w-[400px]', 'p-2'],
+  form: ['flex', 'flex-col', 'w-full', 'items-center', 'gap-2'],
   passwordVisibility: ['w-full', 'max-w-[400px]', 'flex', 'justify-end', 'gap-1'],
+  wrapper: [
+    'flex',
+    'flex-col',
+    'max-w-[600px]',
+    'w-full',
+    'items-center',
+    'gap-[30px]',
+    'p-3',
+    'border',
+    'border-[#252525]/50',
+  ],
+
 };
